:recycle: [REFACTOR]: dto구조 변경
