+++ slash/src/main/java/project/slash/statistics/dto/response/EvaluatedDto.java
@@ -1,4 +1,4 @@
-package project.slash.statistics.dto;
+package project.slash.statistics.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
