+++ slash/src/main/java/project/slash/statistics/repository/StatisticsRepositoryCustom.java
@@ -3,14 +3,14 @@ package project.slash.statistics.repository;
 import java.time.LocalDate;
 import java.util.List;
 
-import project.slash.statistics.dto.MonthlyDataDto;
-import project.slash.statistics.dto.MonthlyServiceStatisticsDto;
-import project.slash.statistics.dto.StatisticsDto;
+import project.slash.statistics.dto.response.MonthlyDataDto;
+import project.slash.statistics.dto.response.MonthlyStatisticsDto;
+import project.slash.statistics.dto.response.StatisticsDto;
 
 public interface StatisticsRepositoryCustom {
 	List<MonthlyDataDto> getMonthlyData(LocalDate date);
 
-	void saveMonthlyData(List<MonthlyServiceStatisticsDto> monthlyServiceStatisticsDtoList);
+	void saveMonthlyData(List<MonthlyStatisticsDto> monthlyStatisticsDtoList);
 
 	List<StatisticsDto> getStatistics(String serviceType, String period, String targetSystem, String targetEquipment);
 }
